wpt_babel/content-security-policy/default-src/default-src-strict_dynamic_and_unsafe_inline.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>
   `strict-dynamic` policy should discard `unsafe-inline` policy.
  </title>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
  <script>
   "use strict";

var t = async_test("Should fire a security policy violation for the inline block");
window.addEventListener('securitypolicyviolation', t.step_func_done(function (e) {
  assert_equals(e.blockedURI, 'inline');
}));
  </script>
  <meta content="default-src 'unsafe-inline' 'strict-dynamic'" http-equiv="Content-Security-Policy"/>
 </head>
 <body>
  <script>
   "use strict";

assert_unreached('Inline script shouldn\'t be run because of the `strict-dynamic` source expression.');
  </script>
 </body>
</html>
